% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSBatchJobsParam-class.R,
%   R/AWSSnowParam-class.R
\name{awsAmiId}
\alias{awsAmiId}
\alias{awsAmiId}
\title{Get AWS AMI-ID of the launched instance. These need to be
Bioconductor configured AMI's.}
\usage{
awsAmiId(x)

awsAmiId(x)
}
\arguments{
\item{AWSBatchJobsParam}{}

\item{AWSSnowParam}{}
}
\description{
Get AWS AMI-ID of the launched instance. These need to be
Bioconductor configured AMI's.

Get AWS AMI-ID of the launched instance
}
