% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSBatchJobsParam-class.R
\docType{class}
\name{AWSBatchJobsParam-class}
\alias{AWSBatchJobsParam-class}
\alias{.AWSBatchJobsParam}
\title{.AWSBatchJobsParam refClass}
\description{
Reference class .AWSBatchJobsParam allows use AWS EC2 as Clusters
}
\details{
The .AWSBatchJobsParam class extends the BatchJobsParam class to allow
usage of AWS EC2-instances for parallel computation. The methods follow a
style similar to that of BiocParallelParams, with bpstart, bpstop, bpisup,
bplapply being the important one. The behaviour of these functions is
described in the man pages.
}
\section{Fields}{

\describe{
\item{\code{awsCredentialsPath}}{Path to AWS credentials, default value is
`~/.aws/credentials`}

\item{\code{awsInstanceType}}{Type of AWS EC2-instance, eg. t2.micro}

\item{\code{awsSubnet}}{AWS EC2-instance subnet, within a certain VPC}

\item{\code{awsAmiId}}{AMI(amazon machine image) ID for the
Bioconductor-starcluster image. Correct ID is needed.}

\item{\code{awsSshKeyPair}}{SSH key pair, to associate with your AWS
EC2-instance}
}}


