% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSParam-class.R
\name{awsCluster}
\alias{awsCluster}
\title{Get the AWSSnowParam object currently launched. Only one AWSSnowParam object can be
started within one session.}
\usage{
awsCluster()
}
\value{
AWSSnowParam object
}
\description{
Get the AWSSnowParam object currently launched. Only one AWSSnowParam object can be
started within one session.
}
