% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSBatchJobsParam-class.R,
%   R/AWSSnowParam-class.R
\name{awsCredentialsPath}
\alias{awsCredentialsPath}
\alias{awsCredentialsPath}
\title{Get path to AWS credentials}
\usage{
awsCredentialsPath(x)

awsCredentialsPath(x)
}
\arguments{
\item{AWSBatchJobsParam}{object}

\item{AWSSnowParam}{object}
}
\description{
Get path to AWS credentials

Get path to AWS credentials
}
