% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.awsDetectMaster}
\alias{.awsDetectMaster}
\title{Function to detect if code is being run on EC2 master node}
\usage{
.awsDetectMaster()
}
\value{
logical
}
\description{
Function to detect if code is being run on EC2 master node
}
\section{Functions}{
\itemize{
\item \code{.awsDetectMaster}: Function to detect if current machine is on
an AWS machine
}}

