% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.awsDetectOrCreateSubnet}
\alias{.awsDetectOrCreateSubnet}
\title{Describe the Subnet that needs to be used}
\usage{
.awsDetectOrCreateSubnet(vpc)
}
\arguments{
\item{vpc}{character subnet is created within the given VPC-ID}
}
\value{
list subnet information
}
\description{
Describe the Subnet that needs to be used
}
\section{Functions}{
\itemize{
\item \code{.awsDetectOrCreateSubnet}: AWS subnet information is detected or created
from given VPC
}}

