% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.awsDetectSecurityGroup}
\alias{.awsDetectSecurityGroup}
\title{Detect the security group which needs to be used}
\usage{
.awsDetectSecurityGroup(vpc)
}
\arguments{
\item{vpc}{character Security Group is created within given VPC-ID}
}
\value{
security group information
}
\description{
Detect the security group which needs to be used
}
\section{Functions}{
\itemize{
\item \code{.awsDetectSecurityGroup}: Detect the security group of the AWS instance
}}

