% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.awsDetectSubnetOnMaster}
\alias{.awsDetectSubnetOnMaster}
\title{Detect the subnet on the AWS master instance of the cluster}
\usage{
.awsDetectSubnetOnMaster()
}
\description{
Detect the subnet on the AWS master instance of the cluster
}
\section{Functions}{
\itemize{
\item \code{.awsDetectSubnetOnMaster}: get the value of subnet on the master node
}}

