% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.awsDetectVpc}
\alias{.awsDetectVpc}
\title{Describe the VPC that needs to be used}
\usage{
.awsDetectVpc(cidr = "10.0.0.0/16")
}
\arguments{
\item{cidr}{character, CIDR block for the VPC}
}
\value{
vpc information
}
\section{Functions}{
\itemize{
\item \code{.awsDetectVpc}: the value for VPC needed to be used
}}

