% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSParam-class.R
\name{awsInstance}
\alias{awsInstance}
\title{Get AWS instance attributes in a list}
\usage{
awsInstance(x)
}
\arguments{
\item{AWSSnowParam}{object}
}
\description{
Get AWS instance attributes in a list
}
