% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSParam-class.R
\name{awsInstanceType}
\alias{awsInstanceType}
\title{Get AWS Instance type.}
\usage{
awsInstanceType(x)
}
\arguments{
\item{AWSSnowParam}{object}
}
\value{
character
}
\description{
The possible instance types are listed in the document:https://aws.amazon.com/ec2/instance-types/. The Bioconductor AMI's have been built using an m4.xlarge instance type. Large computations are best supported on this type of instance.
}
