% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.awsLaunchMaster}
\alias{.awsLaunchMaster}
\title{Launch AWS master node}
\usage{
.awsLaunchMaster(x)
}
\value{
list
}
\description{
Launch AWS master node
}
\section{Functions}{
\itemize{
\item \code{.awsLaunchMaster}: Internal function to launch master node.
}}

