% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSSnowParam-class.R
\name{awsLaunchMasterOnEc2}
\alias{awsLaunchMasterOnEc2}
\title{Launch master node on AWS EC2 if credentials are valid}
\usage{
awsLaunchMasterOnEc2(x)
}
\description{
Launch master node on AWS EC2 if credentials are valid
}
