% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{awsParallelListClusters}
\alias{awsParallelListClusters}
\title{Function to return the names of Clusters launched.}
\usage{
awsParallelListClusters()
}
\description{
Function to return the names of Clusters launched.
}
