% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSParam-class.R
\docType{class}
\name{AWSParam-class}
\alias{AWSParam-class}
\alias{.AWSParam}
\title{Reference class .AWSParam that allows usage of AWS EC2-instances}
\description{
The .AWSParam class extends the BiocParallelParam class
to allow usage of AWS EC2-instances for parallel computation.
The methods follow a style similar to that of BiocParallelParams,
with bpstart, bpstop, bpisup, bplapply being the important one.
}
\section{Fields}{

\describe{
\item{\code{awsCredentialsPath}}{Path to AWS credentials, default value is `~/.aws/credentials`}

\item{\code{awsInstanceType}}{Type of AWS EC2-instance, eg. t2.micro}

\item{\code{awsSubnet}}{AWS EC2-instance subnet, within a certain VPC}

\item{\code{awsSecurityGroup}}{Secutiry group which assigns inbound and outbound traffic at the instance level}

\item{\code{awsInstance}}{A list, created holding all the information of the AWS instance}

\item{\code{awsAmiId}}{AMI(amazon machine image) ID for the Bioconductor-release version}

\item{\code{awsSshKeyPair}}{SSH key pair, to associate with your AWS EC2-instance}
}}


