% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSParam-class.R
\name{AWSParam}
\alias{AWSParam}
\title{AWSParam function to start an AWS EC2-instance cluster}
\usage{
AWSParam(workers = 1, awsCredentialsPath = NA_character_,
  awsInstanceType = NA_character_, awsSubnet = NA, awsSecurityGroup = NA,
  awsAmiId = NA_character_, awsSshKeyPair = NA_character_)
}
\arguments{
\item{workers}{Numeric, number of workers to launch in the cluster}

\item{awsCredentialsPath}{character, Path to AWS credentials, default value is `~/.aws/credentials`}

\item{awsInstanceType}{character, Type of AWS EC2-instance, eg. t2.micro}

\item{awsSubnet}{character, AWS EC2-instance subnet, within a certain VPC}

\item{awsSecurityGroup}{character, Secutiry group which assigns inbound and outbound traffic at the instance level}

\item{awsAmiId}{character, AMI(amazon machine image) ID for the Bioconductor-release version}

\item{awsSshKeyPair}{character, SSH key pair, to associate with your AWS EC2-instance}
}
\description{
This function starts a cluster of AWS EC2-instances to allow
parallel computation of R objects, and works with BiocParallel,
to allow computation with Bioconductor objects
}
