% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSSnowParam-class.R
\name{awsSecurityGroup}
\alias{awsSecurityGroup}
\title{Get AWS Security group for the EC2 instance, which defines inbound and
outbound traffic.}
\usage{
awsSecurityGroup(x)
}
\arguments{
\item{AWSSnowParam}{}
}
\description{
Get AWS Security group for the EC2 instance, which defines inbound and
outbound traffic.
}
