% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSParam-class.R
\name{awsSnowParamCall}
\alias{awsSnowParamCall}
\title{Make socket connection by calling SnowParam}
\usage{
awsSnowParamCall(x)
}
\arguments{
\item{AWSParam}{Object of class AWSParams}
}
\value{
SnowParam object
}
\description{
Make socket connection by calling SnowParam
}
