% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSBatchJobsParam-class.R,
%   R/AWSSnowParam-class.R
\name{awsSshKeyPair}
\alias{awsSshKeyPair}
\alias{awsSshKeyPair}
\title{Get the SSH public key path associted to the AWS EC2 instance.}
\usage{
awsSshKeyPair(x)

awsSshKeyPair(x)
}
\arguments{
\item{AWSBatchJobsParam}{}

\item{AWSSnowParam}{}
}
\description{
Get the SSH public key path associted to the AWS EC2 instance.

Get the SSH public key path associted to the AWS EC2 instance.
}
