% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSBatchJobsParam-class.R,
%   R/AWSSnowParam-class.R
\name{awsSubnet}
\alias{awsSubnet}
\alias{awsSubnet}
\title{Get AWS Subnet within which the AWS EC2 instance was launched}
\usage{
awsSubnet(x)

awsSubnet(x)
}
\arguments{
\item{AWSBatchJobsParam}{}

\item{AWSSnowParam}{}
}
\description{
Get AWS Subnet within which the AWS EC2 instance was launched

Get AWS Subnet within which the AWS EC2 instance was launched
}
