% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSBatchJobsParam-class.R,
%   R/AWSSnowParam-class.R
\name{awsWorkers}
\alias{awsWorkers}
\alias{awsWorkers}
\title{Get number of workers in the cluster}
\usage{
awsWorkers(x)

awsWorkers(x)
}
\arguments{
\item{AWSBatchJobsParam}{object}

\item{AWSSnowParam}{object}
}
\description{
Get number of workers in the cluster

Get number of workers in the cluster
}
