% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSBatchJobsParam-class.R
\name{bpsuspend}
\alias{bpsuspend}
\title{Suspend an AWS EC2 cluster started using bpsetup}
\usage{
bpsuspend(x, clustername = "awsparallel")
}
\arguments{
\item{x}{AWSBatchJobsParam object}

\item{clustername}{character value given to the cluster.}
}
\description{
bpsuspend is required to 'stop' an AWS Cluster, if the user
has an intention of re-using it at a later time. It does NOT
terminate the cluster. The clustername should match the argument
used in bpstart.
}
