% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{.config_starcluster}
\alias{.config_starcluster}
\title{Function to fill the starcluster configuration}
\usage{
.config_starcluster(workers, awsCredentialsPath = "~/.aws/credentials",
  awsInstanceType, awsSubnet = NA_character_, awsAmiId, awsSshKeyPair,
  awsProfile = "default", user = "ubuntu", cidr_ip = NA_character_)
}
\arguments{
\item{workers}{integer vector}

\item{awsCredentialsPath}{character vector}

\item{awsInstanceType}{character vector}

\item{awsSubnet}{character vector}

\item{awsAmiId}{character vector}

\item{awsSshKeyPair}{character vector}

\item{user}{character vector}

\item{cidr_ip}{character vector}
}
\description{
Function to fill the starcluster configuration
}
\section{Functions}{
\itemize{
\item \code{.config_starcluster}: starcluster with the values given in the function
}}

