% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSSnowParam-class.R
\name{getAwsAmiId}
\alias{getAwsAmiId}
\title{Get name of bioconductor release version AMI}
\usage{
getAwsAmiId()
}
\value{
Bioconductor release version
}
\description{
Get name of bioconductor release version AMI
}
