% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getAwsRequirements}
\alias{getAwsRequirements}
\title{Get AWS security requirements}
\usage{
getAwsRequirements()
}
\value{
list, containing VPC, subnet, security group information
}
\description{
Security requirements to launch the EC2 instances into a VPC,
subnet, and security group
}
