% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getStarclusterAmiId}
\alias{getStarclusterAmiId}
\title{This function returns the AMI ID listed on bioconductor
https://www.bioconductor.org/config.yaml}
\usage{
getStarclusterAmiId()
}
\description{
This function returns the AMI ID listed on bioconductor
https://www.bioconductor.org/config.yaml
}
