% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{transferFromCluster}
\alias{transferFromCluster}
\title{Allows transfer of files from master node to the host machine..}
\usage{
transferFromCluster(clustername, remotePath, localPath)
}
\arguments{
\item{clustername}{character vector of the clustername}

\item{starcluster_config}{character vector of path to starcluster config}
}
\description{
Follows this command
    #  starcluster get mycluster --node mycluster-master
    #                            --user myuser /remote/path /local/path
}
