% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{transferToCluster}
\alias{transferToCluster}
\title{Allows transfer of files from Host machine, to master node on cluster.}
\usage{
transferToCluster(clustername, localPath, remotePath)
}
\arguments{
\item{clustername}{character vector of the clustername}

\item{starcluster_config}{character vector of path to starcluster config}
}
\description{
Follows this command
    #  starcluster put mycluster --node mycluster-master
    #                            --user myuser /local/path /remote/path
}
