% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSParam-class.R
\name{awsCluster}
\alias{awsCluster}
\title{Get the AWSParam object currently launched. Only one AWSParam object can be
started within one session.}
\usage{
awsCluster()
}
\value{
AWSParam object
}
\description{
Get the AWSParam object currently launched. Only one AWSParam object can be
started within one session.
}
