% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSParam-class.R
\name{awsCredentialsPath}
\alias{awsCredentialsPath}
\title{Get path to AWS credentials}
\usage{
awsCredentialsPath(x)
}
\arguments{
\item{AWSSnowParam}{object}
}
\description{
Get path to AWS credentials
}
