% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSParam-class.R
\name{awsSshKeyPair}
\alias{awsSshKeyPair}
\title{Get the SSH public key path associted to the AWS EC2 instance.}
\usage{
awsSshKeyPair(x)
}
\arguments{
\item{AWSSnowParam}{}
}
\description{
Get the SSH public key path associted to the AWS EC2 instance.
}
