% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSParam-class.R
\name{awsSubnet}
\alias{awsSubnet}
\title{Get AWS Subnet within which the AWS EC2 instance was launched}
\usage{
awsSubnet(x)
}
\arguments{
\item{AWSParam}{}
}
\description{
Get AWS Subnet within which the AWS EC2 instance was launched
}
