% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWSBatchJobsParam-class.R
\name{bpsetup}
\alias{bpsetup}
\title{Setup cluster where x is clustername}
\usage{
bpsetup(x, clustername = "awsparallel", awsProfile = "default")
}
\description{
Setup cluster where x is clustername
}
